+++ app/safe/internal/state/io-age.go
@@ -1,0 +1,21 @@
+package state
+
+/*
+ * .-'_.---._'-.
+ * ||####|(__)||   Protect your secrets, protect your business.
+ *   \\()|##//       Secure your sensitive data with Aegis.
+ *    \\ |#//                    <aegis.ist>
+ *     .\_/.
+ */
+
+import "strings"
+
+func ageKeyPair() (string, string) {
+	if ageKey == "" {
+		return "", ""
+	}
+
+	parts := strings.Split(ageKey, "\n")
+
+	return parts[0], parts[1]
+}
