+++ app/safe/internal/state/io-encrypt.go
@@ -1,0 +1,43 @@
+/*
+ * .-'_.---._'-.
+ * ||####|(__)||   Protect your secrets, protect your business.
+ *   \\()|##//       Secure your sensitive data with Aegis.
+ *    \\ |#//                    <aegis.ist>
+ *     .\_/.
+ */
+
+package state
+
+import (
+	"filippo.io/age"
+	"github.com/pkg/errors"
+	"github.com/shieldworks/aegis/core/log"
+	"io"
+)
+
+func encryptToWriter(out io.Writer, data string) error {
+	_, publicKey := ageKeyPair()
+	recipient, err := age.ParseX25519Recipient(publicKey)
+	if err != nil {
+		return errors.Wrap(err, "encryptToWriter: failed to parse public key")
+	}
+
+	wrappedWriter, err := age.Encrypt(out, recipient)
+	if err != nil {
+		return errors.Wrap(err, "encryptToWriter: failed to create encrypted file")
+	}
+
+	defer func() {
+		err := wrappedWriter.Close()
+		if err != nil {
+			id := "AEGIIOCL"
+			log.InfoLn(&id, "encryptToWriter: problem closing stream", err.Error())
+		}
+	}()
+
+	if _, err := io.WriteString(wrappedWriter, data); err != nil {
+		return errors.Wrap(err, "encryptToWriter: failed to write to encrypted file")
+	}
+
+	return nil
+}
