+++ app/safe/internal/state/io-persist.go
@@ -1,0 +1,104 @@
+/*
+ * .-'_.---._'-.
+ * ||####|(__)||   Protect your secrets, protect your business.
+ *   \\()|##//       Secure your sensitive data with Aegis.
+ *    \\ |#//                    <aegis.ist>
+ *     .\_/.
+ */
+
+package state
+
+import (
+	"encoding/json"
+	"github.com/pkg/errors"
+	entity "github.com/shieldworks/aegis/core/entity/data/v1"
+	"github.com/shieldworks/aegis/core/env"
+	"github.com/shieldworks/aegis/core/log"
+	"math"
+	"os"
+	"path"
+	"strconv"
+	"time"
+)
+
+var lastBackedUpIndex = make(map[string]int)
+
+func saveSecretToDisk(secret entity.SecretStored, dataPath string) error {
+	data, err := json.Marshal(secret)
+	if err != nil {
+		return errors.Wrap(err, "saveSecretToDisk: failed to marshal secret")
+	}
+
+	file, err := os.Create(dataPath)
+	if err != nil {
+		return errors.Wrap(err, "saveSecretToDisk: failed to create file")
+	}
+	defer func() {
+		err := file.Close()
+		if err != nil {
+			id := "AEGIIOCL"
+			log.InfoLn(&id, "saveSecretToDisk: problem closing file", err.Error())
+		}
+	}()
+
+	return encryptToWriter(file, string(data))
+}
+
+// Only one goroutine accesses this function at any given time.
+func persist(secret entity.SecretStored, errChan chan<- error) {
+	cid := secret.Meta.CorrelationId
+
+	backupCount := env.SafeSecretBackupCount()
+
+	// Resetting the value also removes the secret file from the disk.
+	if secret.Value == "" {
+		dataPath := path.Join(env.SafeDataPath(), secret.Name+".age")
+		err := os.Remove(dataPath)
+		if !os.IsNotExist(err) {
+			log.WarnLn(&cid, "persist: failed to remove secret", err.Error())
+		}
+		return
+	}
+
+	// Save the secret
+	dataPath := path.Join(env.SafeDataPath(), secret.Name+".age")
+
+	err := saveSecretToDisk(secret, dataPath)
+	if err != nil {
+		// Retry once more.
+		time.Sleep(500 * time.Millisecond)
+		err := saveSecretToDisk(secret, dataPath)
+		if err != nil {
+			errChan <- err
+		}
+	}
+
+	index, found := lastBackedUpIndex[secret.Name]
+	if !found {
+		lastBackedUpIndex[secret.Name] = 0
+		index = 0
+	}
+
+	newIndex := math.Mod(float64(index+1), float64(backupCount))
+
+	// Save a copy
+	dataPath = path.Join(
+		env.SafeDataPath(),
+		secret.Name+"-"+strconv.Itoa(int(newIndex))+"-"+".age.backup",
+	)
+
+	err = saveSecretToDisk(secret, dataPath)
+	if err != nil {
+		// Retry once more.
+		time.Sleep(500 * time.Millisecond)
+		err := saveSecretToDisk(secret, dataPath)
+		if err != nil {
+			errChan <- err
+			// Do not change lastBackedUpIndex
+			// since the backup was not successful.
+			return
+		}
+	}
+
+	lastBackedUpIndex[secret.Name] = int(newIndex)
+}
