+++ core/entity/data/v1/v1.go
@@ -9,7 +9,9 @@
 package v1
 
 import (
+	"encoding/json"
 	"fmt"
+	"github.com/shieldworks/aegis/core/template"
 	"time"
 )
 
@@ -74,4 +76,63 @@ type SecretStored struct {
 	// Timestamps
 	Created time.Time
 	Updated time.Time
+}
+
+// ToMapForK8s returns a map that can be used to create a Kubernetes secret.
+//
+//  1. If there is no template, ttempt to unmarshal the secret’ss value
+//     into a map. If that fails, store the secret’s value under the "VALUE" key.
+//  2. If there is a template, attempt to parse it. If parsing is successful,
+//     create a new map with the parsed data. If parsing fails, follow the same
+//     logic as in case 1, attempting to unmarshal the secret’s value into a map,
+//     and if that fails, storing the secret’s value under the "VALUE" key.
+func (secret SecretStored) ToMapForK8s() map[string][]byte {
+	data := make(map[string][]byte)
+
+	// If there is no template, use the secret’s value as is.
+	if secret.Meta.Template == "" {
+		err := json.Unmarshal(([]byte)(secret.Value), &data)
+		if err != nil {
+			value := secret.Value
+			data["VALUE"] = ([]byte)(value)
+		}
+
+		return data
+	}
+
+	// Otherwise, apply the template.
+	newData, err := template.ParseForK8sSecret(secret)
+	if err == nil {
+		data = make(map[string][]byte)
+		for k, v := range newData {
+			data[k] = ([]byte)(v)
+		}
+
+		return data
+	}
+
+	// If the template fails, use the secret’s value as is.
+	err = json.Unmarshal(([]byte)(secret.Value), &data)
+	if err != nil {
+		value := secret.Value
+		data["VALUE"] = ([]byte)(value)
+	}
+
+	return data
+}
+
+// ToMap converts the SecretStored struct to a map[string]any.
+// The resulting map contains the following key-value pairs:
+//
+//	"Name": the Name field of the SecretStored struct
+//	"Value": the Value field of the SecretStored struct
+//	"Created": the Created field of the SecretStored struct
+//	"Updated": the Updated field of the SecretStored struct
+func (secret SecretStored) ToMap() map[string]any {
+	return map[string]any{
+		"Name":    secret.Name,
+		"Value":   secret.Value,
+		"Created": secret.Created,
+		"Updated": secret.Updated,
+	}
 }
