+++ core/probe/privates.go
@@ -1,0 +1,23 @@
+/*
+ * .-'_.---._'-.
+ * ||####|(__)||   Protect your secrets, protect your business.
+ *   \\()|##//       Secure your sensitive data with Aegis.
+ *    \\ |#//                    <aegis.ist>
+ *     .\_/.
+ */
+
+package probe
+
+import (
+	"fmt"
+	"log"
+	"net/http"
+)
+
+func ok(w http.ResponseWriter, _ *http.Request) {
+	_, err := fmt.Fprintf(w, "OK")
+	if err != nil {
+		log.Printf("probe response failure: %s", err.Error())
+		return
+	}
+}
