+++ core/probe/probe.go
@@ -9,20 +9,16 @@
 package probe
 
 import (
-	"fmt"
 	"github.com/shieldworks/aegis/core/env"
 	"log"
 	"net/http"
 )
 
-func ok(w http.ResponseWriter, _ *http.Request) {
-	_, err := fmt.Fprintf(w, "OK")
-	if err != nil {
-		log.Printf("probe response failure: %s", err.Error())
-		return
-	}
-}
-
+// CreateLiveness sets up and starts an HTTP server on the port specified by
+// env.ProbeLivenessPort() to serve as a liveness probe for the application.
+// The server listens for requests at the root path ("/") and responds with an
+// "ok" message. If there is an error starting the server, the function logs
+// a fatal message and returns.
 func CreateLiveness() {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/", ok)
@@ -33,6 +29,11 @@ func CreateLiveness() {
 	}
 }
 
+// CreateReadiness sets up and starts an HTTP server on the port specified by
+// env.ProbeReadinessPort() to serve as a readiness probe for the application.
+// The server listens for requests at the root path ("/") and responds with an
+// "ok" message. If there is an error starting the server, the function logs
+// a fatal message and returns.
 func CreateReadiness() {
 	mux := http.NewServeMux()
 	mux.HandleFunc("/", ok)
